move workers to separate file
